feat: add tsx
