+++ package.json
@@ -3,11 +3,9 @@
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
     "main": "index.js",
-    "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "start": "tsx ./index.ts",
+        "start:dev": "tsx watch ./index.ts"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
@@ -20,9 +18,8 @@
         "eslint": "^8.56.0",
         "eslint-config-prettier": "^9.1.0",
         "eslint-plugin-prettier": "^5.1.3",
-        "nodemon": "^2.0.22",
         "prettier": "^3.2.5",
-        "ts-node": "^10.9.2",
+        "tsx": "^4.7.1",
         "typescript": "^5.3.3"
     },
     "dependencies": {
